Put heading scroll code into its own function
